Modify: query key factory 제거 및 파일 분리
