--- src/apis/channel/useChannelQuery.ts
@@ -1,9 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-
-import channels from "./queryKey";
-
-const useChannelQuery = () => {
-  return useQuery(channels.allChannels);
-};
-
-export default useChannelQuery;
